Move `LexerBaseEx` to `utils` module
