+++ src/main/kotlin/org/rust/lang/core/lexer/RsHighlightingLexer.kt
@@ -6,7 +6,7 @@
 package org.rust.lang.core.lexer
 
 import com.intellij.lexer.LayeredLexer
-import org.rust.lang.core.lexer.RustEscapesLexer.Companion.ESCAPABLE_LITERALS_TOKEN_SET
+import org.rust.lang.core.lexer.RsEscapesLexer.Companion.ESCAPABLE_LITERALS_TOKEN_SET
 import org.rust.lang.core.psi.RS_DOC_COMMENTS
 import org.rust.lang.doc.lexer.RsDocHighlightingLexer
 import org.rust.lang.doc.psi.RsDocKind
@@ -14,7 +14,7 @@ import org.rust.lang.doc.psi.RsDocKind
 class RsHighlightingLexer : LayeredLexer(RsLexer()) {
     init {
         ESCAPABLE_LITERALS_TOKEN_SET.types.forEach {
-            registerLayer(RustEscapesLexer.of(it), it)
+            registerLayer(RsEscapesLexer.of(it), it)
         }
 
         RS_DOC_COMMENTS.types.forEach {
