+++ src/test/kotlin/org/rust/lang/core/lexer/RsEscapesLexingTestCase.kt
@@ -6,7 +6,7 @@
 package org.rust.lang.core.lexer
 
 import com.intellij.lexer.Lexer
-import org.rust.lang.core.lexer.RustEscapesLexer.Companion.ESCAPABLE_LITERALS_TOKEN_SET
+import org.rust.lang.core.lexer.RsEscapesLexer.Companion.ESCAPABLE_LITERALS_TOKEN_SET
 import org.rust.lang.core.psi.RsElementTypes.BYTE_STRING_LITERAL
 import org.rust.lang.core.psi.RsElementTypes.STRING_LITERAL
 import java.util.*
@@ -15,11 +15,11 @@ class RsEscapesLexingTestCase : RsLexingTestCaseBase() {
     override fun getTestDataPath(): String = "org/rust/lang/core/lexer/fixtures/escapes"
 
     // We want unicode and eol but not extended escapes. String literals are perfect.
-    override fun createLexer(): Lexer = RustEscapesLexer.of(STRING_LITERAL)
+    override fun createLexer(): Lexer = RsEscapesLexer.of(STRING_LITERAL)
 
     fun testShortByteEscapes() = doTest()
     fun testLongByteEscapes() = doTest()
-    fun testExtendedByteEscapes() = doTest(RustEscapesLexer.of(BYTE_STRING_LITERAL))
+    fun testExtendedByteEscapes() = doTest(RsEscapesLexer.of(BYTE_STRING_LITERAL))
     fun testSampleUnicodeEscapes() = doTest()
     fun testVariableLengthUnicodeEscapes() = doTest()
     fun testInvalidShortByteEscapes() = doTest()
@@ -30,7 +30,7 @@ class RsEscapesLexingTestCase : RsLexingTestCaseBase() {
     fun testUnderscoreInUnicodeEscapes() = doTest()
 
     fun testFuzzy() {
-        val lexers = ESCAPABLE_LITERALS_TOKEN_SET.types.map { RustEscapesLexer.of(it) }
+        val lexers = ESCAPABLE_LITERALS_TOKEN_SET.types.map { RsEscapesLexer.of(it) }
         for (lexer in lexers) {
             repeat(10000) {
                 randomLiteral().tokenize(lexer).forEach { }
